float16 resolved, GetZeroValue unresovled
